Added isMobile in the utils index file and hence refactored a little.
