+++ app/ui-sidenav/client/sidebarItem.js
@@ -3,13 +3,12 @@ import { ReactiveVar } from 'meteor/reactive-var';
 import { Session } from 'meteor/session';
 import { Template } from 'meteor/templating';
 
-import { t, getUserPreference, roomTypes } from '../../utils';
+import { t, getUserPreference, roomTypes, isMobile } from '../../utils';
 import { popover, renderMessageBody, menu } from '../../ui-utils';
 import { Users, ChatSubscription } from '../../models';
 import { settings } from '../../settings';
 import { hasAtLeastOnePermission } from '../../authorization';
 import { timeAgo } from '../../lib/client/lib/formatDate';
-import { isMobile } from '../../utils/client/lib/isMobile';
 
 Template.sidebarItem.helpers({
 	streaming() {
